refactor: options cleanup
Mostly just move the code.
